+++ pil2-components/lib/std/rs/src/range_check/u16air.rs
@@ -98,7 +98,7 @@ impl U16Air {
             let row_idx = value & self.mask;
 
             // Update the multiplicity
-            self.multiplicities[range_idx][row_idx].fetch_add(*multiplicity as u64, Ordering::Relaxed);
+            self.multiplicities[range_idx][row_idx].fetch_add(*multiplicity, Ordering::Relaxed);
         }
     }
